+++ Changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Latest
+* Added waypoint 1 task copying for steersman
 ## 1.5.4a
 * Added meters for short distances in hitchtrooper sitrep
 ## 1.5.3a