+++ Changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Latest
+* Added evac for hitchtroopers without an evac point set
+* Added one-time hitchtroopers that can be activated by mission triggers rather than the comms menus
+* Added instruction hints in the hitchtroopers 'Available' comms item
+##1.5.5a
 * Added waypoint 1 task copying for steersman
 * Added map mark generation for hitch troopers
 * Added default frequencies for asset pool groups